change: moved about page to home page, removed about route, and updated the header and footer nav links
