+++ next.config.mjs
@@ -1,20 +1,11 @@
-import rehypePrism from '@mapbox/rehype-prism'
-import nextMDX from '@next/mdx'
-import remarkGfm from 'remark-gfm'
+import rehypePrism from "@mapbox/rehype-prism";
+import nextMDX from "@next/mdx";
+import remarkGfm from "remark-gfm";
 
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-  pageExtensions: ['js', 'jsx', 'ts', 'tsx', 'mdx'],
-  async redirects() {
-    return [
-      {
-        source: '/',
-        destination: '/about',
-        permanent: true,
-      },
-    ]
-  },
-}
+  pageExtensions: ["js", "jsx", "ts", "tsx", "mdx"],
+};
 
 const withMDX = nextMDX({
   extension: /\.mdx?$/,
@@ -22,6 +13,6 @@ const withMDX = nextMDX({
     remarkPlugins: [remarkGfm],
     rehypePlugins: [rehypePrism],
   },
-})
+});
 
-export default withMDX(nextConfig)
+export default withMDX(nextConfig);
