+++ src/app/layout.tsx
@@ -1,29 +1,28 @@
-import { type Metadata } from 'next'
+import { type Metadata } from "next";
 
-import { Providers } from '@/app/providers'
-import { Layout } from '@/components/Layout'
+import { Providers } from "@/app/providers";
+import { Layout } from "@/components/Layout";
 
-import '@/styles/tailwind.css'
+import "@/styles/tailwind.css";
 
 export const metadata: Metadata = {
   title: {
-    template: '%s - Spencer Sharp',
-    default:
-      'Spencer Sharp - Software designer, founder, and amateur astronaut',
+    template: "%s | Mike Frieze",
+    default: "Mike Frieze",
   },
   description:
-    'I’m Spencer, a software designer and entrepreneur based in New York City. I’m the founder and CEO of Planetaria, where we develop technologies that empower regular people to explore space on their own terms.',
+    "Hi, I’m Mike Frieze. I live in Detroit, where I develop UIs for the web.",
   alternates: {
     types: {
-      'application/rss+xml': `${process.env.NEXT_PUBLIC_SITE_URL}/feed.xml`,
+      "application/rss+xml": `${process.env.NEXT_PUBLIC_SITE_URL}/feed.xml`,
     },
   },
-}
+};
 
 export default function RootLayout({
   children,
 }: {
-  children: React.ReactNode
+  children: React.ReactNode;
 }) {
   return (
     <html lang="en" className="h-full antialiased" suppressHydrationWarning>
@@ -35,5 +34,5 @@ export default function RootLayout({
         </Providers>
       </body>
     </html>
-  )
+  );
 }
